auth_service/requirements.txt
authlib
fastapi
passlib[bcrypt]
pymysql
python-dotenv
python-jose[cryptography]
python-multipart
sqlalchemy
uvicorn
